requirements.txt
ollama>=0.4.0
PyQt5>=5.15.0
Pillow>=10.0.0
